requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
pandas>=1.2,<2.0
numpy>=1.21,<2.0
scikit-learn>=0.24,<1.0
python-dateutil>=2.8,<3.0
pytz>=2024.0
scipy>=1.5,<2.0
joblib>=1.0,<2.0